+++ scripts/data/exchange/index.js
@@ -1,5 +1,18 @@
-const ews = require('./ews')
+const getRoomAvailability = require('./room-availability')
+const getEmails = require('./get-unique-emails')
+const mapRoomsToTimes = require('./rooms-to-times')
+const getOpeningHours = require('./opening-hours')
+const removeProps = require('./remove-props')
 
-module.exports = {
-  addOpeningHours: ews
+require('dotenv').config()
+
+const useMockData = process.env.USE_MOCK_DATA_EXCHANGE === '1'
+
+module.exports = async ({ spaces, buildings } = {}) => {
+  const emails = getEmails(spaces)
+  const availability = await getRoomAvailability(emails, useMockData)
+  const roomsToTimesMap = mapRoomsToTimes(emails, availability)
+  const openingHours = getOpeningHours(roomsToTimesMap, { rooms: spaces, buildings })
+  const cleanedBuildingsAndSpaces = removeProps({ buildings, spaces })
+  return { ...cleanedBuildingsAndSpaces, openingHours }
 }
