+++ scripts/data/index.js
@@ -2,7 +2,7 @@ require('dotenv').config()
 
 const { writeFile } = require('fs')
 const getDataFromCms = require('./cms')
-const { addOpeningHours } = require('./exchange')
+const addOpeningHours = require('./exchange')
 const validate = require('./validate')
 
 const { CSV_PATH: csvPath } = process.env
